Move breadcrumb logic to `index.tsx` file
